+++ gprof/getnfile.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -488,7 +489,8 @@ unsigned long strsize)
     unsigned long i;
 
 	for(i = 0; i < nsymbols; i++){
-	    if(symbols[i].n_un.n_strx != 0 && symbols[i].n_un.n_strx < strsize){
+	    if(symbols[i].n_un.n_strx != 0 &&
+	       (unsigned long)symbols[i].n_un.n_strx < strsize){
 		if(funcsymbol(symbols + i, strings + symbols[i].n_un.n_strx))
 		    nname++;
 	    }
@@ -507,7 +509,8 @@ unsigned long vmaddr_slide)
     unsigned long i;
 
 	for(i = 0; i < nsymbols; i++){
-	    if(symbols[i].n_un.n_strx != 0 && symbols[i].n_un.n_strx < strsize){
+	    if(symbols[i].n_un.n_strx != 0 &&
+	       (unsigned long)symbols[i].n_un.n_strx < strsize){
 		if(funcsymbol(symbols + i, strings + symbols[i].n_un.n_strx)){
 		    npe->value = symbols[i].n_value + vmaddr_slide;
 		    npe->name = strings + symbols[i].n_un.n_strx;
@@ -581,7 +584,7 @@ unsigned long strsize)
 	    if(symbols[i].n_type == N_SO){
 		/* skip the N_SO for the directory name that ends in a '/' */
 		if(symbols[i].n_un.n_strx != 0 &&
-		   symbols[i].n_un.n_strx < strsize){
+		   (unsigned long)symbols[i].n_un.n_strx < strsize){
 		    name = strings + symbols[i].n_un.n_strx;
 		    len = strlen(name);
 		    if(len != 0 && name[len-1] == '/')
@@ -610,7 +613,7 @@ unsigned long strsize)
 	    if(symbols[i].n_type == N_SO){
 		/* skip the N_SO for the directory name that ends in a '/' */
 		if(symbols[i].n_un.n_strx != 0 &&
-		   symbols[i].n_un.n_strx < strsize){
+		   (unsigned long)symbols[i].n_un.n_strx < strsize){
 		    name = strings + symbols[i].n_un.n_strx;
 		    len = strlen(name);
 		    if(len != 0 && name[len-1] == '/')
