+++ gprof/printgprof.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -118,7 +119,7 @@ void)
 {
     nltype *np;
     nltype **sortednlp;
-    int index;
+    unsigned long index;
 
 	actime = 0.0;
 	printf("\f\n");
@@ -281,7 +282,7 @@ void
 printgprof(
 nltype **timesortnlp)
 {
-    int index;
+    unsigned long index;
     nltype *parentp;
 
 	/*
@@ -824,7 +825,7 @@ void)
 {
     nltype **namesortnlp;
     nltype *nlp;
-    int index, nnames, todo, i, j;
+    unsigned long index, nnames, todo, i, j;
     char peterbuffer[BUFSIZ];
 
 	/*
@@ -842,7 +843,7 @@ void)
 	}
 	qsort(namesortnlp, nnames, sizeof(nltype *),
 	      (int (*)(const void *, const void *))namecmp);
-	for(index = 1, todo = nnames; index <= ncycle; index++){
+	for(index = 1, todo = nnames; index <= (unsigned long)ncycle; index++){
 	    namesortnlp[todo++] = &cyclenl[index];
 	}
 	printf("\f\nIndex by function name\n\n");
