+++ ld/rld.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -612,7 +613,7 @@ long obj_size)
 		 * Write the entire output file.
 		 */
 		if(write(fd, output_addr, output_size + symbol_size) !=
-		   output_size + symbol_size){
+		   (int)(output_size + symbol_size)){
 		    system_error("can't write output file: %s",output_filename);
 #ifdef KLD
 		    internal_kld_unload(TRUE);
@@ -824,7 +825,7 @@ const char *base_filename,
 char *base_addr,
 long base_size)
 {
-    long size;
+    unsigned long size;
     char *addr;
     int fd;
     struct stat stat_buf;
@@ -976,7 +977,7 @@ long base_size)
 #endif /* __LITTLE_ENDIAN__ */
     
 		if(sizeof(struct fat_header) + fat_header->nfat_arch *
-		sizeof(struct fat_arch) > size){
+		   sizeof(struct fat_arch) > (unsigned long)size){
 		    error("fat file: %s truncated or malformed (fat_arch "
 			"structs would extend past the end of the file)",
 			base_name);
@@ -1638,7 +1639,7 @@ const char *output_filename)
 	     * Write the entire output file.
 	     */
 	    if(write(fd, output_addr, output_size + symbol_size) !=
-	       output_size + symbol_size){
+	       (int)(output_size + symbol_size)){
 		system_error("can't write output file: %s",output_filename);
 		(void)unlink(output_filename);
 		return_value = 0;
