+++ libstuff/writeout.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -511,7 +512,7 @@ unsigned long *throttle)
 		    write_size = (file + file_size) - p;
 		else
 		    write_size = WRITE_SIZE;
-		if(write(fd, p, write_size) != write_size){
+		if(write(fd, p, write_size) != (int)write_size){
 		    system_error("can't write output file: %s", output);
 		    goto cleanup;
 		}
@@ -560,7 +561,7 @@ unsigned long *throttle)
 	}
 	else{
 no_throttle:
-	    if(write(fd, file, file_size) != file_size){
+	    if(write(fd, file, file_size) != (int)file_size){
 		system_error("can't write output file: %s", output);
 		goto cleanup;
 	    }
@@ -807,7 +808,7 @@ enum bool library_warnings)
 		    strings_size = object->output_strings_size;
 		}
 		for(j = 0; j < nsymbols; j++){
-		    if(symbols[j].n_un.n_strx > strings_size)
+		    if((unsigned long)symbols[j].n_un.n_strx > strings_size)
 			continue;
 		    if(toc_symbol(symbols + j, commons_in_toc,
 		       object->sections) == TRUE){
@@ -998,7 +999,7 @@ struct arch *arch,
 char *output,
 enum bool library_warnings)
 {
-    long i;
+    unsigned long i;
     enum bool multiple_defs;
     struct member *member;
 
@@ -1008,7 +1009,7 @@ enum bool library_warnings)
 	 * only once (marked by changing the sign of their ran_off).
 	 */
 	multiple_defs = FALSE;
-	for(i = 0; i < (long)arch->toc_nranlibs - 1; i++){
+	for(i = 0; i < arch->toc_nranlibs - 1; i++){
 	    if(strcmp(arch->toc_ranlibs[i].ran_un.ran_name,
 		      arch->toc_ranlibs[i+1].ran_un.ran_name) == 0){
 		if(multiple_defs == FALSE){
